AUTH-1480 - Move code validation to static helper
- To remove the need of creating an instance of the ValidationService
